Add IQ test and category tabs, update activities logic
- Added TabControl to the Activities tab, including a "Health and Wellness" category.
- Converted `Player.Smarts` to `double` and updated UI to show formatted percentages.
- Added restrictions allowing activities to be performed only once per year.
- Restricted the Gym activity to players 12 years and older.
- Added new activities: Dieting, Gardening, Visiting the Library, Meditating, and Go for a Walk.
- Implemented a multi-level Math IQ test that increases Smarts by 0.25 (up to 5%) per correct question and fails instantly on incorrect choices.

+++ Player.cs
@@ -9,7 +9,7 @@ namespace BitLifeClone
         public int Age { get; set; }
         public int Health { get; set; }
         public int Happiness { get; set; }
-        public int Smarts { get; set; }
+        public double Smarts { get; set; }
         public int Looks { get; set; }
         public decimal Money { get; set; }
         public string CurrentEducation { get; set; } = "None";
